Send json over websockets and display
need to come up with a workflow for defining, and changing the object, then sending on bang. This should be simple but requires a little UX design I think. Could also do with some tidying up, but with the package.json (and auto file tidying turned off in max), things are grand!